pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "bootstrapper"
version = "0.2.0"
description = "Bootstrap volume segmentations from sparse annotations"
readme = "README.md"
authors = [
    {name = "Vijay Venu Thiyagarajan", email = "vvenu@utexas.edu"}
]
license = {text = "MIT"}
dependencies = [
    "numpy",
    "scipy",
    "scikit-image",
    "torch",
    "zarr",
    "numba",
    "neuroglancer",
    "tensorboardx",
    "gunpowder>=1.4",
    "daisy>=1.2",
    "lsds @ git+https://github.com/yajivunev/lsd.git",
    "funlib.learn.torch @ git+https://github.com/funkelab/funlib.learn.torch.git",
    "funlib.show.neuroglancer @ git+https://github.com/funkelab/funlib.show.neuroglancer.git",
    "funlib.segment @ git+https://github.com/funkelab/funlib.segment.git",
    "funlib.evaluate @ git+https://github.com/yajivunev/funlib.evaluate.git",
    "waterz @ git+https://github.com/mpinb/waterz.git",
    "click",
]

[project.optional-dependencies]
mws = [
    "mwatershed @ git+https://github.com/pattonw/mwatershed.git",
    "volara @ git+https://github.com/pattonw/volara.git",
]
all = ["bootstrapper[mws]"]

[project.urls]
Homepage = "https://github.com/ucsdmanorlab/bootstrapper"

[project.scripts]
bs = "bootstrapper.cli:cli"